Add Panic Handling to listusers expandDirect
